refactoring.

+++ database/odbc/detail/get_data.hpp
@@ -11,7 +11,7 @@ namespace brig { namespace database { namespace odbc { namespace detail {
 
 struct get_data : boost::noncopyable {
   virtual ~get_data()  {}
-  virtual SQLRETURN operator()(SQLHSTMT stmt, int col, variant& var) = 0;
+  virtual SQLRETURN operator()(SQLHSTMT stmt, size_t col, variant& var) = 0;
 }; // get_data
 
 } } } } // brig::database::odbc::detail
